// src/app/main-layout/table-data.interface.ts
export interface TableDataInterface {
  name: string;
  position: number;
  status: string;
  weight: number;
  symbol: string;
  ownerStatus?: string;
  tenant?: string

}
/*
created: string;
property: string;
propertyStatus: string;
plan: string;
owner: string;
ownerStatus: string;
tenant: string
*/